backend/pyproject.toml
[project]
name = "Rental-Backend"
version = "0.1.0"
description = "Application for rental application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.6",
    "django-admin>=2.0.2",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.15.2",
    "isort>=6.0.1",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
]

